+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/dao/preset/ProviderInfoPresetCriteriaMaker.java
@@ -13,12 +13,14 @@ import java.util.Objects;
 @Component
 public class ProviderInfoPresetCriteriaMaker implements PresetCriteriaMaker {
 
-    @SuppressWarnings("SwitchStatementWithTooFewBranches")
     @Override
     public void makeCriteria(DetachedCriteria criteria, String preset, Object[] objs) {
         switch (preset) {
             case ProviderInfoMaintainService.CHILD_FOR_STATISTICS_SETTING:
                 childForStatisticsSetting(criteria, objs);
+                break;
+            case ProviderInfoMaintainService.CHILD_FOR_STATISTICS_SETTING_ENABLED:
+                childForStatisticsSettingEnabled(criteria, objs);
                 break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + preset);
@@ -34,6 +36,21 @@ public class ProviderInfoPresetCriteriaMaker implements PresetCriteriaMaker {
                 LongIdKey longIdKey = (LongIdKey) objs[0];
                 criteria.add(Restrictions.eqOrIsNull("statisticsSettingLongId", longIdKey.getLongId()));
             }
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objs));
+        }
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    private void childForStatisticsSettingEnabled(DetachedCriteria criteria, Object[] objs) {
+        try {
+            if (Objects.isNull(objs[0])) {
+                criteria.add(Restrictions.isNull("statisticsSettingLongId"));
+            } else {
+                LongIdKey longIdKey = (LongIdKey) objs[0];
+                criteria.add(Restrictions.eqOrIsNull("statisticsSettingLongId", longIdKey.getLongId()));
+            }
+            criteria.add(Restrictions.eq("enabled", true));
         } catch (Exception e) {
             throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objs));
         }