+++ statistics-stack/src/main/java/com/dwarfeng/statistics/stack/service/DriverInfoMaintainService.java
@@ -16,4 +16,5 @@ public interface DriverInfoMaintainService extends BatchCrudService<LongIdKey, D
         EntireLookupService<DriverInfo>, PresetLookupService<DriverInfo> {
 
     String CHILD_FOR_STATISTICS_SETTING = "child_for_statistics_setting";
+    String CHILD_FOR_STATISTICS_SETTING_ENABLED = "child_for_statistics_setting_enabled";
 }